Added function to display frames
Experimented with Blessed. I'm keeping that experiment in case the program changes, since it allows us to look at and copy simpler code if we need to. Added makeFrame( ) which will be passed a 2D array - like the playing grid - and display it in the terminal. Updated documentation with details about it.